rename: MenuBarIcons.tsx → MenuItemIcons.tsx
