+++ components/menu-icons/index.ts
@@ -1,4 +1,4 @@
-export { MenuItemIcon } from "./MenuBarIcons";
+export { MenuItemIcon } from "./MenuItemIcons";
 export { HomeIcon } from "./HomeIcon";
 export { MenuIcon } from "./MenuIcon";
 export { OrderIcon } from "./OrderIcon";
